academia/sig788/code/task_7_HD_Sarang_Manohars223504903_bot_v1/requirements.txt
aiohttp
botbuilder-integration-aiohttp
botbuilder-core
botbuilder-schema
cookiecutter==1.7.0
azure-core
azure-ai-language-conversations
google-api-python-client